Update/move GW unbalanced test
